Create new file structure
