+++ pypaperless/models/base.py
@@ -1,11 +1,11 @@
 """Provide base classes."""
 
 from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Any, ClassVar, Protocol, Self, TypeVar, final
+from typing import TYPE_CHECKING, Any, ClassVar, Self, TypeVar, final
 
 from pydantic import BaseModel, ConfigDict, PrivateAttr
 
-from pypaperless.const import API_PATH, PaperlessResource
+from pypaperless.const import API_PATH
 
 if TYPE_CHECKING:
     from pydantic import TypeAdapter
@@ -14,31 +14,6 @@ if TYPE_CHECKING:
 
 
 ResourceT = TypeVar("ResourceT", bound="PaperlessModel")
-
-
-class PaperlessBase:
-    """Superclass for all classes in PyPaperless."""
-
-    _api_path = API_PATH["index"]
-
-    def __init__(self, client: "Paperless") -> None:
-        """Initialize a `PaperlessBase` instance."""
-        self._client = client
-
-
-class ServiceProtocol[ResourceT](Protocol):
-    """Protocol for any `ServiceBase` instances and its ancestors."""
-
-    _client: "Paperless"
-    _api_path: str
-    _resource: PaperlessResource
-    _resource_cls: type[ResourceT]
-
-
-class ServiceBase(PaperlessBase):
-    """Base class for all services in PyPaperless."""
-
-    _resource: PaperlessResource
 
 
 class PaperlessModel(BaseModel):
