+++ pypaperless/models/config.py
@@ -1,11 +1,10 @@
-"""Provide `Config` related models and services."""
+"""Provide `Config` related models."""
 
 from typing import TYPE_CHECKING, Any, ClassVar
 
-from pypaperless.const import API_PATH, PaperlessResource
+from pypaperless.const import API_PATH
 
-from .base import ServiceBase, PaperlessModel
-from .mixins import services
+from .base import PaperlessModel
 
 if TYPE_CHECKING:
     from pypaperless import Paperless
@@ -48,12 +47,3 @@ class Config(PaperlessModel):
         """Initialize a `Config` instance."""
         super().__init__(client, data, **kwargs)
         self._format_api_path(data)
-
-
-class ConfigService(ServiceBase, services.CallableMixin[Config]):
-    """Represent a factory for Paperless `Config` models."""
-
-    _api_path = API_PATH["config"]
-    _resource = PaperlessResource.CONFIG
-
-    _resource_cls = Config
