--- pypaperless/models/mixins/models/__init__.py
@@ -1,12 +1,0 @@
-"""Mixins for PyPaperless models."""
-
-from .creatable import CreatableMixin
-from .data_fields import MatchingFieldsMixin
-from .securable import SecurableDraftMixin, SecurableMixin
-
-__all__ = (
-    "CreatableMixin",
-    "MatchingFieldsMixin",
-    "SecurableDraftMixin",
-    "SecurableMixin",
-)
