--- pypaperless/models/mixins/models/deletable.py
@@ -1,4 +1,0 @@
-"""DeletableMixin has been moved to services.
-
-See pypaperless.models.mixins.services.deletable for the service-side implementation.
-"""
