--- pypaperless/models/mixins/models/updatable.py
@@ -1,4 +1,0 @@
-"""UpdatableMixin has been moved to services.
-
-See pypaperless.models.mixins.services.updatable for the service-side implementation.
-"""
