+++ pypaperless/models/statistics.py
@@ -1,10 +1,10 @@
-"""Provide `Statistics` related models and services."""
+"""Provide `Statistics` related models."""
 
 from typing import ClassVar
 
-from pypaperless.const import API_PATH, PaperlessResource
+from pypaperless.const import API_PATH
 
-from .base import PaperlessModel, ServiceBase
+from .base import PaperlessModel
 from .common import StatisticDocumentFileTypeCount
 
 
@@ -24,17 +24,3 @@ class Statistic(PaperlessModel):
     document_type_count: int | None = None
     storage_path_count: int | None = None
     current_asn: int | None = None
-
-
-class StatisticService(ServiceBase):
-    """Represent a factory for Paperless `Statistic` models."""
-
-    _api_path = API_PATH["statistics"]
-    _resource = PaperlessResource.STATISTICS
-
-    _resource_cls = Statistic
-
-    async def __call__(self) -> Statistic:
-        """Request the `Statistic` model data."""
-        res = await self._client.request_json("get", self._api_path)
-        return self._resource_cls.create_with_data(self._client, res)
