+++ pypaperless/models/status.py
@@ -1,8 +1,8 @@
-"""Provide `Status` related models and services."""
+"""Provide `Status` related models."""
 
 from typing import ClassVar, cast
 
-from pypaperless.const import API_PATH, PaperlessResource
+from pypaperless.const import API_PATH
 from pypaperless.models.common import (
     StatusDatabaseType,
     StatusStorageType,
@@ -10,7 +10,7 @@ from pypaperless.models.common import (
     StatusType,
 )
 
-from .base import PaperlessModel, ServiceBase
+from .base import PaperlessModel
 
 
 class Status(PaperlessModel):
@@ -42,17 +42,3 @@ class Status(PaperlessModel):
         ]
 
         return any(st == StatusType.ERROR for st in statuses)
-
-
-class StatusService(ServiceBase):
-    """Represent a factory for the Paperless `Status` model."""
-
-    _api_path = API_PATH["status"]
-    _resource = PaperlessResource.STATUS
-
-    _resource_cls = Status
-
-    async def __call__(self) -> Status:
-        """Request the `Status` model data."""
-        res = await self._client.request_json("get", self._api_path)
-        return self._resource_cls.create_with_data(self._client, res)
