+++ pypaperless/models/tags.py
@@ -1,0 +1,62 @@
+"""Provide `Tag` related models."""
+
+from typing import TYPE_CHECKING, Any, ClassVar
+
+from pypaperless.const import API_PATH
+
+from . import mixins
+from .base import PaperlessModel
+
+if TYPE_CHECKING:
+    from pypaperless import Paperless
+
+
+class Tag(
+    PaperlessModel,
+    mixins.MatchingFieldsMixin,
+    mixins.SecurableMixin,
+):
+    """Represent a Paperless `Tag`."""
+
+    _api_path: ClassVar[str] = API_PATH["tags_single"]
+
+    id: int | None = None
+    slug: str | None = None
+    name: str | None = None
+    color: str | None = None
+    text_color: str | None = None
+    is_inbox_tag: bool | None = None
+    document_count: int | None = None
+    parent: int | None = None
+    children: list["Tag"] | None = None
+
+    def __init__(self, client: "Paperless", data: dict[str, Any], **kwargs: Any) -> None:
+        """Initialize a `Tag` instance."""
+        super().__init__(client, data, **kwargs)
+        self._format_api_path(data)
+
+
+class TagDraft(
+    PaperlessModel,
+    mixins.MatchingFieldsMixin,
+    mixins.SecurableDraftMixin,
+    mixins.CreatableMixin,
+):
+    """Represent a new `Tag`, which is not yet stored in Paperless."""
+
+    _api_path: ClassVar[str] = API_PATH["tags"]
+
+    _create_required_fields: ClassVar[set[str]] = {
+        "name",
+        "color",
+        "is_inbox_tag",
+        "match",
+        "matching_algorithm",
+        "is_insensitive",
+    }
+
+    name: str | None = None
+    color: str | None = None
+    text_color: str | None = None
+    is_inbox_tag: bool | None = None
+    owner: int | None = None
