+++ pypaperless/services/mixins/callable.py
@@ -2,7 +2,8 @@
 
 from typing import Any
 
-from pypaperless.models.base import ResourceT, ServiceProtocol
+from pypaperless.models.base import ResourceT
+from pypaperless.services.base import ServiceProtocol
 
 
 class CallableMixin(ServiceProtocol[ResourceT]):
